(AB-1951706) Simplify PR Template (#64)
Prior to this commit, the PR Template was a combined kitchen-sink checklist
and guidance document to help a contributor describe their pull request and
declare they had followed the guidance.

This commit simplifies the template down to a summary and checklist that is
universal for all PRs filed against the repository. The checklist now asks
only that the contributor has written a descriptive title and summary, that
they have read the contributors guide, and that they have adhered to the
style guide.

This simplified template is intended for use with several in-development
GitHub Actions to replace functionality previously captured in the template,
namely:

- The Expectations GHA will comment on new PRs by community members to let
  them know more specifically what to expect from the process.
- The Authorization checks (live and path) will provide on-PR feedback when
  a contributor is targeting files or a branch when they are not authorized
  to do so.
- The Versioned Content Change Report will provide on-PR feedback when authors
  change versioned content, showing what was changed across versions.
- The Checklist GHA will review a PR that is not in draft or marked WIP to
  verify that the checklist in this simplified template has been filled out.

Future GHAs may enhance the PR experience further.