(AB-4580) Add PSDscResources reference docs (#86)
* (AB-4580) Add PSDscResources reference docs

Prior to this change, the **PSDscResources** module and its resources were
only documented in GitHub. With many of the in-box DSC Resources incompatible
with DSC v2, referencing the more up-to-date and supported **PSDscResources**
module in the v2 docs is preferable.

This change pulls over that documentation and updates it for readability,
maintainability, accuracy, and our documentation standards.

This change includes examples for each resource from the GitHub repo. Each
example is in a separate file with a short description and uses code snippet
syntax to display the example as both a Configuration definition and as a
script using Invoke-DscResource, since Configurations are no longer directly
applicable.

Aside from the maintenance concerns listed above, this change does not alter
the documentation. It may be beneficial to rewrite the examples into short
tutorials or otherwise readdress this domain.

TODO:

- Determine a sensible TOC for this
- Update metadata as required

This change fixes AB#4580.

* Add subtoc and update toc

* Fix typo

* (MAINT) Reformat code fences

* (MAINT) Rearrange TOC

This change ensures the resource is the first item in the TOC for each
subfolder.

* (MAINT) Minor reorg of base TOC

This change moves the PSDscResource module docs into a section for
DSC resource module reference

Co-authored-by: Sean Wheeler <sean.wheeler@microsoft.com>