(AB-4757) Add terminology for DSC
Prior to this change, the terminology for DSC used a mix of casing and
was not always clear because the terms `Configuration`, `Resource`, and
`Node`, in particular, are overloaded.

This change adds a new document to list and briefly explain terms used
for DSC. This document should form the basis of how these concepts are
written in the documentation and all formatting and usage should be
updated to comply with the standards set here.