Add documentation for v2 (#94)
* (AB-1962930) Scaffold v2 documentation (#72)

* (AB-1967391) Draft Getting Started for v2 (#74)

Prior to this change, the "Get started with invoking DSC Resources"
article was merely a direct copy of the "Calling DSC resource methods
directly" article from the v1.1 docs.

This change:

- Rewrites the document for getting started with DSC v2 via the
  `Invoke-DscResource` cmdlet.
- Deletes the unnecessary windows-specific getting started document.
- Resolves AB#1967391

* (AB#4242) Remove references to Pull Server for v2 (#76)

Prior to this change, documentation for v2 referenced the Pull Server,
which is removed in v2.

This change:

- Removes direct references to the Pull Server
- Resolves AB#4242

* (AB#4301) Purge removed cmdlets (#82)

Prior to this change the v2 docs referenced cmdlets
and behaviors that no longer exist in v2. This change
addresses that problem by removing those references
and rewriting the documentation as required.

Resolves AB#4301

* (AB#4589) Improve DSC v2 maintainability (#84)

* (AB-4580) Add PSDscResources reference docs (#86)

* (AB-4580) Add PSDscResources reference docs

Prior to this change, the **PSDscResources** module and its resources were
only documented in GitHub. With many of the in-box DSC Resources incompatible
with DSC v2, referencing the more up-to-date and supported **PSDscResources**
module in the v2 docs is preferable.

This change pulls over that documentation and updates it for readability,
maintainability, accuracy, and our documentation standards.

This change includes examples for each resource from the GitHub repo. Each
example is in a separate file with a short description and uses code snippet
syntax to display the example as both a Configuration definition and as a
script using Invoke-DscResource, since Configurations are no longer directly
applicable.

Aside from the maintenance concerns listed above, this change does not alter
the documentation. It may be beneficial to rewrite the examples into short
tutorials or otherwise readdress this domain.

TODO:

- Determine a sensible TOC for this
- Update metadata as required

This change fixes AB#4580.

* Add subtoc and update toc

* Fix typo

* (MAINT) Reformat code fences

* (MAINT) Rearrange TOC

This change ensures the resource is the first item in the TOC for each
subfolder.

* (MAINT) Minor reorg of base TOC

This change moves the PSDscResource module docs into a section for
DSC resource module reference

Co-authored-by: Sean Wheeler <sean.wheeler@microsoft.com>

* (AB-4757) Add terminology for DSC (#87)

* (AB-4757) Add terminology for DSC

Prior to this change, the terminology for DSC used a mix of casing and
was not always clear because the terms `Configuration`, `Resource`, and
`Node`, in particular, are overloaded.

This change adds a new document to list and briefly explain terms used
for DSC. This document should form the basis of how these concepts are
written in the documentation and all formatting and usage should be
updated to comply with the standards set here.

* Pair review of PR

Co-authored-by: Sean Wheeler <sean.wheeler@microsoft.com>

* (AB-4757) Standardize and limit docs (#91)

This changes standardizes the v2 documentation to use the terms in the
new terminology document. It also removes extraneous documentation that
doesn't apply to Azure Policy's machine configuration feature.

This change:

- Removes the docs for dynamic DSC Configurations and for Configuration
  Data as they do not apply when using DSC with machine policy.
- Removes the troubleshooting documentation drafts until they can be
  fully written.
- Standardizes terminology and formatting.
- Fixes AB#4757

* (AB-4581) Add PSDscResource link to v1.1 (#92)

This change adds an entry to the v1.1 TOC to link readers to
the PSDscResources documentation in the v2 docs.

- Fixes AB#4581

* (AB-5011) Reorganize v2 docs TOC (#93)

This change completely reorganizes the TOC for the DSC v2 documentation
to make the information easier to discover, navigate, and maintain.

This change:

- Replaces the prior structure, splitting the documentation primarily
  between the `Configurations` and `Resources` nodes, with `Concepts`
  and `How-to Guides`.

  This change moves the following documents to the `Concepts` node:

  - Terminology
  - DSC Resources
  - Get, Test, and Set
  - DSC Configurations
  - Import-DSCResource

  This change moves the following documents to the `How-to guides` node:

  - Installing DSC Resources
  - Authoring DSC Resources (node group)
    - Overview
    - Authoring a class-based DSC Resource
    - Authoring a single-instance DSC Resource
    - Authoring a MOF-based DSC Resource
    - Authoring a Composite DSC Resource
    - DSC Resource authoring checklist
  - Write and compile a DSC Configuration
  - Define DSC Resource dependencies in a DSC Configuration using
    DependsOn

  The rest of the documentation structure remains unchanged.
- Updates links in moved/renamed files to reflect the correct structure
  and makes minor updates/adjustments for clarity.
- Resolves AB#5011

Co-authored-by: Sean Wheeler <sean.wheeler@microsoft.com>