(AB-5074) Update markdownlint config to yaml
This change replaces the JSON configuration for markdownlint in the
repository with YAML files, which can be commented and are easier
to author and maintain.

The config for the DSC Resources in v2 was removed because the updated
config, based on the one in MicrosoftDocs/PowerShell-Docs, sets MD024
(`no-duplicate-header`) to only check for siblings - headers with the
same name in the same section.

- Related to AB#5074