(AB#5132) Add cSpell configuration
This change adds configuration for cSpell based on the configuration
added in MicrosoftDocs/PowerShell-Docs#9170. The `psdocs` configuration
is a copy from that repository, while the scaffolded `dsc` configuration
is specific to this one. The JSON configuration in the root of the
`.vscode` folder imports both definitions, merging them for checking the
spelling of Markdown documents in this repository.

The `dsc` dictionary is currently empty as no terms needed to be added.

The `psdocs` dictionary in this repository should never be edited except
to keep it in sync with the definition in MicrosoftDocs/PowerShell-Docs.