(AB#5227) Address machine config limits for DSC Resources
There is a known limitation in machine configuration which prevents
users from using DSC Resources which call any cmdlets which are not
included in PowerShell itself or available from a module on the Gallery.
This includes the built-in Windows modules, like **DISM**.

This change updates the documentation to:

- Alert readers to this limitation where sensible
- Replace examples in the conceptual docs which point to DSC Resources
  that won't work in machine configuration with ones that will
- Ensure all conceptual docs which apply only to machine configuration
  say so at the top of the document.

This change fixes AB#5227.