(AB#5178) Add class-based DSC Resource concept
Prior to this change, the only document with detail on the
implementation of class-based DSC Resources was the how-to guide, which
tried to enumerate all of the structural concerns of a class-based
resource at the same time as it showed a user how to write one.

Reader feedback indicated that this was confusing, difficult, and
required a lot of investigation to understand how a class-based DSC
Resource works and should be written.

This change introduces a new conceptual document for class-based DSC
Resources to fully document what they are, how they're structured,
and concerns for authoring them.

With this document added, the how-to guide can be rewritten and
simplified so it takes a user from start to finish for authoring a basic
class-based DSC Resource. A second how-to guide should be added for
authoring class-based DSC Resources for Azure Policy's machine
configuration feature so readers can get specific information about the
context they're writing their DSC Resource for.

Fixes AB#5178