(MAINT) Cleanup metadata and "applies to"
This change removes the version-specific annotation at the top of the
documentation specific to DSC Resources in the v2 docs - these documents
are not actually version-dependent, since the DSC Resources are valid in
Windows PowerShell and with DSC v1.1.

It also corrects the version-specific annotation for DSC Configuration
documents to specify PowerShell 7 instead of 7.2, as DSC v2 actually
works in older versions of PowerShell (though not Windows PowerShell) -
it only stopped _shipping_ in PowerShell in 7.2.

Finally, this commit reorders the metadata alphabetically for all docs
in the v2 docset for maintainability.