(MAINT) Document behavior for value & ref types
Prior to this change, the documentation did not mention the differences
between reference and value types as they apply to DSC Resources.

This change adds the _Default property values_ section, which explains
the intialization of a DSC Resource's properties. Particularly, it
covers their default values and how that differs between reference types
and value types.

It clarifies the impact of this, namely that a DSC Resource can't know
with certainty whether a value-type properties was specified as its
default value or not specified at all and therefore can't be written
to treat those properties as unmanaged if their default value is a valid
value for that property.

It also updates the _Use enums instead of ValidateSet_ section to call
out the impact this has on enums and advises readers accordingly.

Finally, it adds two new best practices sections for writing DSC
Resources with the default values for reference and value type
properties in mind.