(GH-126) Update docs for Reasons property
This change updates the documentation across the versions for the
**Reasons** property required by Azure Automanage's machine
configuration feature.

Prior to this change, the documentation showed creating a class for the
**Reasons** complex property called `Reason`. While this works, it also
leads DSC Resource authors towards implementing the class with that name,
which leads to errors when compiling a configuration that uses two or
more modules that define a class with the same name.

This change:

- Updates the how-to documentation for authoring a class-based DSC
  Resource across every version to use a prefix before the class name
  for the complex property and note why the prefix is used
- Add a new section to the conceptual documentation for class-based DSC
  Resources in v2 to explain the property's usage and requirements.
- Resolves #126