(AB#5183) Add tutorial for machine config
Prior to this change, the only tutorial for writing a class-based DSC
Resource was for resources not intended for use with the machine
configuration feature of Azure Automanage.

This change adds a specific authoring tutorial for that use case,
including the usage of the **Reasons** property for reporting how the
DSC Resource is out of the desired state.

- Fixes AB#5183