(AB-5183) Address PR Feedback
This change updates the implementation for the DSC Resource
to return Reasons for properties even when they are in the
desired state, per feedback.

It reimplements the `Test()` method bringing it back in line
with the other class-based DSC Resource implementation now
that the `$Reasons` property being populated is not a proxy
for whether the DSC Resource is in the desired state.