(GH-147) Fix v3 getting started examples
Prior to this change, the example code for the v3 getting started
article showed passing the parameters for `Invoke-DscResource` as
a hash directly to the command.

This change fixes the examples by storing the parameter values as
a hashtable in a variable and then splatting the hash table with
the cmdlet execution.

- Fixes AB#82869
- Resolves #147