(AB#118304) Add reference docs for command-based resources
This change adds reference documentation for the proof-of-concept
prototype command-based DSC Resources `Microsoft/OSInfo` and
`Microsoft.Windows/Registry`.

The reference documentation proposes a structure for the reference
documentation for command-based resources based on the reference
documentation used for the PSDscResources in v2.

It also documents the command-line tools the resources use.

While these commands and resources are not production-ready, they are
the only command-based resources we can reference in the documentation
at this time. If the resources are removed or replaced in the future,
the documentation can be updated.

This change:

- Fixes AB#118304
- Fixes AB#118305
- Fixes AB#118306