Draft resource authoring design guide
This change drafts the conceptual documentation for
guidance on how to author DSC resources from a
design perspective. It does not include any specific
guidance from a code or language perspective.

Instead, it focuses on the high level concepts and on
defining the resource manifest and instance JSON Schema.

This change is a mirror of the one submitted to the
source repository:

- PowerShell/DSC#1027